feat: Implement sending and receiving SSE toasts.
